Refactor: move around text selection code
