+++ crates/egui/src/text_selection/mod.rs
@@ -1,0 +1,5 @@
+//! Helpers regarding text selection for labels and text edit.
+
+mod label_text_selection;
+
+pub use label_text_selection::label_text_selection;
